demo for order process

+++ Papara-Final-Project/Papara-Final-Project/DTOs/UserDTO.cs
@@ -8,7 +8,7 @@
         public string Email { get; set; }
         public string Role { get; set; }
         public decimal WalletBalance { get; set; }
-        public int PointsBalance { get; set; }
+        public decimal PointsBalance { get; set; }
         public string Token { get; set; }
     }
 }
